feat: proposal hurrying
Added capability to hurry proposal. Only available for OwnerRole. As for now, only the DAO itself would be able to use this method. This means it should go through another proposal, which kind of defeats the purpose.

Still, I added it to future proof the governance capability. A new component could be made that is only allowed to call the hurry_vote method, and is controlled by some kind of Governance Observer / Administrator.