fix: bootstrap store historic state in ledger
The bootstrap component was wrongly configured to overwrite previous historic state when state was updated, meaning only 1 state was ever stored.

This fix ensures all state changes are stored appropriately (after each swap the progress and both token supplies are stored).

This enables proper graphing of the price history of the bootstrapped token.